Move WorkshopRequestExternalForm and tests to extforms application
