+++ amy/workshops/tests/test_workshop_requests.py
@@ -1,7 +1,8 @@
-from django.core import mail
 from django.urls import reverse
 
-from workshops.tests.base import TestBase
+from workshops.forms import (
+    WorkshopRequestBaseForm,
+)
 from workshops.models import (
     WorkshopRequest,
     Event,
@@ -12,11 +13,7 @@ from workshops.models import (
     ComputingExperienceLevel,
     Curriculum,
 )
-from ..forms import (
-    WorkshopRequestBaseForm,
-    WorkshopRequestAdminForm,
-    WorkshopRequestExternalForm,
-)
+from workshops.tests.base import TestBase
 
 
 class TestWorkshopRequestBaseForm(TestBase):
@@ -163,7 +160,7 @@ class TestWorkshopRequestBaseForm(TestBase):
         # 1: valid self-organized
         data = {
             'organization_type': 'self',
-            'self_organized_github': \
+            'self_organized_github':
                 'http://hogwarts.github.io/2018-11-03-Hogwarts',
             'centrally_organized_fee': '',
             'waiver_circumstances': '',
@@ -298,7 +295,6 @@ class TestWorkshopRequestBaseForm(TestBase):
     def test_travel_expences_management(self):
         """Test validation of travel expences management."""
 
-
         # 1: data required
         data = {
             'travel_expences_management': '',
@@ -328,90 +324,13 @@ class TestWorkshopRequestBaseForm(TestBase):
         self.assertNotIn('travel_expences_management_other', form.errors)
 
 
-class TestWorkshopRequestExternalForm(TestBase):
-    """Test external (accessible to non-logged in users) form."""
-
-    def test_fields_presence(self):
-        """Test if the form shows correct fields."""
-        form = WorkshopRequestExternalForm()
-        fields_left = set(form.fields.keys())
-        fields_right = set([
-            "personal", "family", "email", "institution", "institution_name",
-            "institution_department", "location", "country",
-            "part_of_conference", "conference_details", "preferred_dates",
-            "language", "number_attendees", "domains", "domains_other",
-            "academic_levels", "computing_levels", "audience_description",
-            "requested_workshop_types", "organization_type",
-            "self_organized_github", "centrally_organized_fee",
-            "waiver_circumstances", "travel_expences_agreement",
-            "travel_expences_management", "travel_expences_management_other",
-            "comment", "data_privacy_agreement", "code_of_conduct_agreement",
-            "host_responsibilities", "captcha",
-        ])
-        self.assertEqual(fields_left, fields_right)
-
-    def test_request_added(self):
-        """Ensure the request is successfully added to the pool, and
-        notification email is sent."""
-        data = {
-            'personal': 'Harry',
-            'family': 'Potter',
-            'email': 'hpotter@magic.gov',
-            'institution_name': 'Ministry of Magic',
-            'location': 'London',
-            'country': 'GB',
-            'preferred_dates': '03-04 November, 2018',
-            'language':  Language.objects.get(name='English').pk,
-            'number_attendees': '10-40',
-            'domains': [],
-            'domains_other': 'Wizardry',
-            'academic_levels': [AcademicLevel.objects.first().pk],
-            'computing_levels': [ComputingExperienceLevel.objects.first().pk],
-            'audience_description': 'Students of Hogwarts',
-            'requested_workshop_types': [
-                Curriculum.objects.get(slug='swc-python').pk,
-                Curriculum.objects.get(slug='dc-ecology-r').pk,
-            ],
-            'organization_type': 'central',
-            'self_organized_github': '',
-            'centrally_organized_fee': 'waiver',
-            'waiver_circumstances': 'Bugdet cuts in Ministry of Magic',
-            'travel_expences_agreement': True,
-            'travel_expences_management': 'booked',
-            'travel_expences_management_other': '',
-            'comment': 'N/c',
-            'data_privacy_agreement': True,
-            'code_of_conduct_agreement': True,
-            'host_responsibilities': True,
-        }
-        self.passCaptcha(data)
-
-        rv = self.client.post(reverse('workshop_request'), data,
-                              follow=True)
-        self.assertEqual(rv.status_code, 200)
-        content = rv.content.decode('utf-8')
-        if 'form' in rv.context:
-            self.assertEqual(rv.context['form'].is_valid(), True,
-                             dict(rv.context['form'].errors))
-        self.assertNotIn('Please fix errors in the form below', content)
-        self.assertIn('Thank you for requesting a workshop', content)
-        self.assertEqual(WorkshopRequest.objects.all().count(), 1)
-        self.assertEqual(WorkshopRequest.objects.all()[0].state, 'p')
-        self.assertEqual(len(mail.outbox), 1)
-        msg = mail.outbox[0]
-        self.assertEqual(
-            msg.subject,
-            'New workshop request: Ministry of Magic, 03-04 November, 2018',
-        )
-        self.assertEqual(msg.recipients(), ['admin-uk@carpentries.org'])
-
-
 class TestWorkshopRequestViews(TestBase):
     def setUp(self):
         self._setUpUsersAndLogin()
 
         self.wr1 = WorkshopRequest.objects.create(
-            state="p", personal="Harry", family="Potter", email="harry@potter.com",
+            state="p", personal="Harry", family="Potter",
+            email="harry@potter.com",
             institution_name="Hogwarts", location="Scotland", country="GB",
             part_of_conference=False, preferred_dates="soon",
             language=Language.objects.get(name='English'),
@@ -419,7 +338,8 @@ class TestWorkshopRequestViews(TestBase):
             organization_type='self',
         )
         self.wr2 = WorkshopRequest.objects.create(
-            state="d", personal="Harry", family="Potter", email="harry@potter.com",
+            state="d", personal="Harry", family="Potter",
+            email="harry@potter.com",
             institution_name="Hogwarts", location="Scotland", country="GB",
             part_of_conference=False, preferred_dates="soon",
             language=Language.objects.get(name='English'),
@@ -493,7 +413,7 @@ class TestWorkshopRequestViews(TestBase):
     def test_discarded_request_reopened(self):
         self.wr1.state = "a"
         self.wr1.save()
-        rv = self.client.get(
+        self.client.get(
             reverse('workshoprequest_set_state',
                     args=[self.wr1.pk, 'pending']),
             follow=True)
@@ -502,7 +422,7 @@ class TestWorkshopRequestViews(TestBase):
 
     def test_accepted_request_reopened(self):
         self.assertEqual(self.wr2.state, "d")
-        rv = self.client.get(
+        self.client.get(
             reverse('workshoprequest_set_state',
                     args=[self.wr2.pk, 'pending']),
             follow=True)
